refactor: Splitting repo into lib and bin
