+++ Cargo.toml
@@ -1,7 +1,14 @@
 [package]
-name = "shuftle-client"
+name = "shuftle-client-core"
 version = "0.1.0"
 edition = "2021"
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "shuftle-client"
+path = "src/main.rs"
 
 [dependencies]
 bevy = { version = "0.14.1", features = ["dynamic_linking"] }
