+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod camera;
+pub mod load_assets;
+pub mod main_player;
